+++ examples/basic/clean.sh
@@ -1,0 +1,6 @@
+set -e
+
+buildDir=dist
+
+echo "Removing $buildDir"
+rm -rf $buildDir