+++ examples/standalone/check.sh
@@ -1,0 +1,6 @@
+set -e
+
+pnpm test
+pnpm libmodulor GenerateAppsTests
+pnpm libmodulor TestApp --appName MyRunningMap
+pnpm libmodulor TestApp --appName Toolbox