+++ examples/standalone/src/apps/MyRunningMap/src/lib/link/FakeLinkManager.ts
@@ -1,0 +1,23 @@
+import { injectable } from 'inversify';
+import { sleep } from 'libmodulor';
+
+import type {
+    LinkManager,
+    LinkManagerOpenable,
+    LinkManagerOpenOpts,
+} from './LinkManager.js';
+
+@injectable()
+export class FakeLinkManager implements LinkManager {
+    public history: [LinkManagerOpenable, LinkManagerOpenOpts | undefined][] =
+        [];
+
+    public async open(
+        openable: LinkManagerOpenable,
+        opts?: LinkManagerOpenOpts,
+    ): Promise<void> {
+        this.history.push([openable, opts]);
+        await sleep(300);
+        opts?.onClose?.(openable);
+    }
+}