+++ examples/standalone/src/apps/MyRunningMap/src/ucds/GenerateRunningMapFromStravaUCD.ts
@@ -40,7 +40,7 @@ export interface GenerateRunningMapFromStravaOPI0 extends UCOPIBase {
 }
 
 @injectable()
-export class GenerateRunningMapFromStravaClientMain
+class GenerateRunningMapFromStravaClientMain
     implements
         UCMain<
             GenerateRunningMapFromStravaInput,