+++ examples/standalone/src/apps/Toolbox/src/lib/std/FakeLLMManager.ts
@@ -1,0 +1,21 @@
+import { injectable } from 'inversify';
+import type {
+    LLMManager,
+    LLMManagerSendReq,
+    LLMManagerSendRes,
+} from 'libmodulor';
+
+@injectable()
+export class FakeLLMManager implements LLMManager {
+    public async send(req: LLMManagerSendReq): Promise<LLMManagerSendRes> {
+        return {
+            choices: [
+                {
+                    message: {
+                        content: `I'm not able to process your request : ${JSON.stringify(req)}`,
+                    },
+                },
+            ],
+        };
+    }
+}