+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Misc**
 
 - Add the `Toolbox` app to `examples` (for docs)
+- Expose `UC_POLICY_FILE_NAME_EXT` and `UC_POLICY_FILE_NAME_SUFFIX` in the convention
 
 ## v0.19.0 (2025-06-20)